IOS- Sui: fix codestyle
