+++ BlockchainSdk/Blockchains/Sui/SuiAddressService.swift
@@ -10,7 +10,6 @@ import Foundation
 import WalletCore
 
 final class SuiAddressService: AddressService {
-    
     func makeAddress(for publicKey: Wallet.PublicKey, with addressType: AddressType) throws -> any Address {
         try WalletCoreAddressService(coin: .sui).makeAddress(for: publicKey, with: .default)
     }
@@ -18,5 +17,4 @@ final class SuiAddressService: AddressService {
     func validate(_ address: String) -> Bool {
         WalletCoreAddressService(coin: .sui).validate(address)
     }
-    
 }
