+++ BlockchainSdk/Blockchains/Sui/SuiWalletAssembly.swift
@@ -8,9 +8,7 @@
 
 import Foundation
 
-
 struct SuiWalletAssembly: WalletManagerAssembly {
-    
     func make(with input: WalletManagerAssemblyInput) throws -> any WalletManager {
         let providers = APIResolver(blockchain: input.blockchain, config: input.blockchainSdkConfig)
             .resolveProviders(apiInfos: input.apiInfo) { nodeInfo, _ in
@@ -22,5 +20,4 @@ struct SuiWalletAssembly: WalletManagerAssembly {
         
         return SuiWalletManager(wallet: input.wallet, networkService: SuiNetworkService(providers: providers))
     }
-    
 }
